services/control-plane/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
PyJWT==2.10.1
webauthn==2.7.0
